requirements.txt
# NAAC Compliance Intelligence System Dependencies

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.8.2
pydantic-settings==2.2.1

# RAG Components
# Align sentence-transformers with latest huggingface_hub to avoid cached_download import errors
sentence-transformers==2.7.0

# LLM Integration
groq>=0.18.0
huggingface-hub==0.33.4
requests==2.31.0

# Document Processing
pdfplumber==0.10.3
pypdf2==3.0.1
pdfminer.six==20221105
python-multipart==0.0.6

# Vector math helpers
numpy==2.1.3
psycopg2-binary==2.9.9

# Scheduling (removed scheduler feature; keep dependency list clean)

# Web Scraping for Updates
beautifulsoup4==4.12.2

# Environment Management
python-dotenv==1.0.0

# Testing
pytest==8.2.2

# NOTE:
# sqlite3, pathlib, and hashlib are Python standard-library modules.
# They must not be listed in requirements.txt.